requirements.txt
google-genai>=1.0.0
python-dotenv>=0.21.0
opencv-python>=4.0.0
numpy>=1.20.0
pillow>=9.0.0